Parsed file structure
